+++ src/analysis/mod.rs
@@ -1,3 +1,8 @@
+//! Analysis pass of the parser
+//!
+//! This is just if for some reason you want to split the parsing from the
+//! analysis.
+
 use crate::error::PassResult;
 use crate::ScalableRecipe;
 
@@ -8,7 +13,7 @@ pub use event_consumer::parse_events;
 pub type AnalysisResult = PassResult<ScalableRecipe>;
 
 #[derive(PartialEq, Eq, Debug, Clone, Copy)]
-pub enum DefineMode {
+pub(crate) enum DefineMode {
     All,
     Components,
     Steps,
@@ -16,7 +21,7 @@ pub enum DefineMode {
 }
 
 #[derive(PartialEq, Eq, Debug, Clone, Copy)]
-pub enum DuplicateMode {
+pub(crate) enum DuplicateMode {
     New,
     Reference,
 }